Refactor internal contract
