+++ core/src/executive/internal_contract/components/activation.rs
