+++ core/src/executive/internal_contract/contracts/pos.rs
@@ -2,17 +2,9 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-use super::{
-    super::impls::pos::*, macros::*, SimpleExecutionTrait, SolFnTable,
-};
-use crate::{
-    evm::{ActionParams, Spec},
-    executive::InternalRefContext,
-    observer::VmObserve,
-    vm,
-};
+use super::{super::impls::pos::*, preludes::*};
+
 use cfx_parameters::internal_contract_addresses::POS_REGISTER_CONTRACT_ADDRESS;
-use cfx_state::state_trait::StateOpsTrait;
 use cfx_types::{Address, H256, U256};
 
 type Bytes = Vec<u8>;
