+++ core/src/executive/internal_contract/contracts/sponsor.rs
@@ -2,15 +2,7 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-use super::{
-    super::impls::sponsor::*, macros::*, SimpleExecutionTrait, SolFnTable,
-};
-use crate::{
-    evm::{ActionParams, Spec},
-    executive::InternalRefContext,
-    observer::VmObserve,
-    vm,
-};
+use super::{super::impls::sponsor::*, preludes::*};
 use cfx_parameters::internal_contract_addresses::SPONSOR_WHITELIST_CONTROL_CONTRACT_ADDRESS;
 use cfx_types::{Address, U256};
