+++ core/src/executive/internal_contract/impls/mod.rs
@@ -5,7 +5,7 @@
 pub(super) mod admin;
 pub(super) mod cross_space;
 pub(super) mod params_control;
-pub(crate) mod pos;
+pub(super) mod pos;
 pub(super) mod sponsor;
 pub(super) mod staking;
